feat: refactor and add dropdown variable component
