+++ libs/domains/variables/feature/src/index.ts
@@ -11,6 +11,6 @@ export * from './lib/variable-list/variable-list'
 export * from './lib/create-update-variable-modal/create-update-variable-modal'
 export * from './lib/variables-context/variables-context'
 export * from './lib/show-all-variables-toggle/show-all-variables-toggle'
-export * from './lib/code-editor-variables/code-editor-variables'
+export * from './lib/code-editor-variable/code-editor-variable'
 export * from './lib/flow-create-variable/flow-create-variable'
 export * from './lib/field-variable-suggestion/field-variable-suggestion'
