--- libs/domains/variables/feature/src/lib/code-editor-variables/code-editor-variables.spec.tsx
@@ -1,9 +1,0 @@
-import { renderWithProviders } from '@qovery/shared/util-tests'
-import { CodeEditorVariables } from './code-editor-variables'
-
-describe('CodeEditorVariables', () => {
-  it('should render successfully', () => {
-    const { baseElement } = renderWithProviders(<CodeEditorVariables environmentId="000" />)
-    expect(baseElement).toBeTruthy()
-  })
-})
